requirements.txt
numpy 
onnxruntime
opencv-python
pydantic-settings
filterpy
faiss-cpu
ultralight